+++ install.sh
@@ -226,7 +226,7 @@ get_ffmpeg_configure_opts() {
 --disable-inline-asm
 --disable-doc
 --enable-muxer=flac,mov,ipod,wav,w64,ffmetadata
---enable-demuxer=flac,mov,wav,w64,aiff,ffmetadata,dsf,aac,hls,mpegts,mp3,ogg,pcm_s16le,pcm_s24le,pcm_s32le,pcm_f32le,lavfi
+--enable-demuxer=flac,mov,wav,w64,aiff,ffmetadata,dsf,aac,hls,mpegts,mp3,ogg,pcm_s16le,pcm_s16be,pcm_s24le,pcm_s24be,pcm_s32le,pcm_s32be,pcm_f32le,lavfi
 --enable-encoder=alac,flac,pcm_s16le,pcm_s24le,pcm_s32le
 --enable-decoder=alac,flac,pcm_s16le,pcm_s24le,pcm_s32le,pcm_f32le,pcm_s16be,pcm_s24be,pcm_s32be,dsd_lsbf,dsd_msbf,dsd_lsbf_planar,dsd_msbf_planar,vorbis,aac,aac_fixed,aac_latm,mp3,mp3float,mjpeg,png
 --enable-parser=aac,aac_latm,flac,vorbis,mpegaudio,mjpeg
@@ -262,7 +262,7 @@ get_ffmpeg_8_minimal_opts() {
 --disable-avdevice
 --disable-swscale
 --enable-protocol=file,http,https,tcp,hls
---enable-demuxer=flac,wav,aiff,dsf,aac,mov,mp3,ogg,hls,mpegts
+--enable-demuxer=flac,wav,aiff,dsf,aac,mov,mp3,ogg,hls,mpegts,pcm_s16be
 --enable-decoder=flac,alac,pcm_s16le,pcm_s24le,pcm_s32le,pcm_f32le,pcm_s16be,pcm_s24be,pcm_s32be,dsd_lsbf,dsd_msbf,dsd_lsbf_planar,dsd_msbf_planar,aac,aac_fixed,aac_latm,mp3,mp3float,vorbis
 --enable-parser=aac,aac_latm,mpegaudio,vorbis
 --enable-muxer=flac,wav
@@ -1360,7 +1360,7 @@ WRAPPER_EOF
         fi
 
         # Migrate settings from old config
-        local KNOWN_KEYS="TARGET PORT NAME RENDERER_NAME GAPLESS VERBOSE MINIMAL_UPNP INTERFACE NETWORK_INTERFACE THREAD_MODE CYCLE_TIME CYCLE_MIN_TIME INFO_CYCLE TRANSFER_MODE TARGET_PROFILE_LIMIT MTU MTU_OVERRIDE CPU_AUDIO CPU_OTHER NICE_LEVEL IO_SCHED_CLASS IO_SCHED_PRIORITY RT_PRIORITY"
+        local KNOWN_KEYS="TARGET PORT NAME RENDERER_NAME GAPLESS VERBOSE MINIMAL_UPNP INTERFACE NETWORK_INTERFACE THREAD_MODE CYCLE_TIME CYCLE_MIN_TIME INFO_CYCLE TRANSFER_MODE TARGET_PROFILE_LIMIT MTU MTU_OVERRIDE CPU_AUDIO CPU_OTHER PCM_BUFFER_SECONDS PCM_REMOTE_BUFFER_SECONDS DSD_BUFFER_SECONDS PCM_PREFILL_MS PCM_REMOTE_PREFILL_MS DSD_PREFILL_MS NICE_LEVEL IO_SCHED_CLASS IO_SCHED_PRIORITY RT_PRIORITY"
         local migrated_keys=""
         local obsolete_keys=""